Gtk3: split mixer widgets into VSID/non-VSID
The code controlling the layout was getting too complex. VSID now has its own
mixer widget while the other emus have theirs, for use in both the settings UI
and as a statusbar widget.
The colons after the labels in the CRT controls have been removed to keep a
consistent look.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@34933 379a1393-f5fb-40a0-bcee-ef074d9b53f7