Gtk3: Make Gtk3 UI use the new CBM font
Update buildsystem to use and install the new font, update archdep code to use the new font.
Looks like the indexes of the chars in the new font are different from the old font, so the conversion function needs to be updated. It also looks like the new font has different padding from the old font, so the CSS we use to create proper dirart seems off.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@37470 379a1393-f5fb-40a0-bcee-ef074d9b53f7