Rename PET ROMs to include their part number (in line with Zimmers).
This should make it more clear which editor ROMs exist. While doing this, I
discovered that the 4032B model used a wrong Editor (a version without CRTC
support), so I changed it to a different one. I also added 60 Hz versions, even
though the GUI does not yet have a quick way to select a model that uses them.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38230 379a1393-f5fb-40a0-bcee-ef074d9b53f7